requirements.txt
Flask
Flask-SQLAlchemy
paho-mqtt
Adafruit-DHT
